+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Use new `ProtectedHeader` type for protected headers (breaking change).  This variant of `Header` preserves any
   originally-parsed data, so that calculations (signatures, decryption, etc.) over the data can use the bit-for-bit wire
   data instead of a reconstituted (and potentially different) version.
+- Use new `OutOfRangeIntegerValue` error when an integer value is too large for the representation
+  used in this crate (breaking change).
 
 ## 0.2.0 - 2021-12-09